pyproject.toml
[project]
name = "safiro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "adrf>=0.1.11",
    "colorlog>=6.10.1",
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "drf-spectacular-sidecar>=2025.10.1",
    "httpx>=0.28.1",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.1.0",
    "drf-spectacular-extras>=0.1.0",
    "ipdb>=0.13.13",
]
